fix(cache): address p3 review feedback
