+++ crates/rginx-config/src/compile/tests/cache.rs
@@ -342,271 +342,3 @@ fn compile_cache_policy_supports_p0_controls() {
     assert_eq!(policy.lock_timeout, Duration::from_secs(2));
     assert_eq!(policy.lock_age, Duration::from_secs(3));
 }
-
-#[test]
-fn compile_cache_policy_supports_disabling_p2_defaults() {
-    let base_dir = temp_base_dir("rginx-cache-compile-p2-disable");
-    let config = Config {
-        cache_zones: vec![CacheZoneConfig {
-            name: "default".to_string(),
-            path: "cache/default".to_string(),
-            max_size_bytes: Some(1024 * 1024),
-            inactive_secs: Some(120),
-            default_ttl_secs: Some(30),
-            max_entry_bytes: Some(1024),
-            path_levels: None,
-            loader_batch_entries: None,
-            loader_sleep_millis: None,
-            manager_batch_entries: None,
-            manager_sleep_millis: None,
-            inactive_cleanup_interval_secs: None,
-            shared_index: Some(false),
-        }],
-        runtime: RuntimeConfig {
-            shutdown_timeout_secs: 10,
-            worker_threads: None,
-            accept_workers: None,
-        },
-        listeners: Vec::new(),
-        server: ServerConfig {
-            listen: Some("127.0.0.1:8080".to_string()),
-            server_header: None,
-            proxy_protocol: None,
-            default_certificate: None,
-            server_names: Vec::new(),
-            trusted_proxies: Vec::new(),
-            client_ip_header: None,
-            keep_alive: None,
-            max_headers: None,
-            max_request_body_bytes: None,
-            max_connections: None,
-            header_read_timeout_secs: None,
-            request_body_read_timeout_secs: None,
-            response_write_timeout_secs: None,
-            access_log_format: None,
-            tls: None,
-            http3: None,
-        },
-        upstreams: vec![UpstreamConfig {
-            name: "backend".to_string(),
-            peers: vec![UpstreamPeerConfig {
-                url: "http://127.0.0.1:9000".to_string(),
-                weight: 1,
-                backup: false,
-            }],
-            tls: None,
-            dns: None,
-            protocol: UpstreamProtocolConfig::Auto,
-            load_balance: UpstreamLoadBalanceConfig::RoundRobin,
-            server_name: None,
-            server_name_override: None,
-            request_timeout_secs: None,
-            connect_timeout_secs: None,
-            read_timeout_secs: None,
-            write_timeout_secs: None,
-            idle_timeout_secs: None,
-            pool_idle_timeout_secs: None,
-            pool_max_idle_per_host: None,
-            tcp_keepalive_secs: None,
-            tcp_nodelay: None,
-            http2_keep_alive_interval_secs: None,
-            http2_keep_alive_timeout_secs: None,
-            http2_keep_alive_while_idle: None,
-            max_replayable_request_body_bytes: None,
-            unhealthy_after_failures: None,
-            unhealthy_cooldown_secs: None,
-            health_check_path: None,
-            health_check_grpc_service: None,
-            health_check_interval_secs: None,
-            health_check_timeout_secs: None,
-            healthy_successes_required: None,
-        }],
-        locations: vec![LocationConfig {
-            cache: Some(CacheRouteConfig {
-                zone: "default".to_string(),
-                methods: Some(vec!["GET".to_string(), "HEAD".to_string()]),
-                statuses: Some(vec![200]),
-                ttl_secs_by_status: None,
-                key: Some("{scheme}:{host}:{uri}".to_string()),
-                cache_bypass: None,
-                no_cache: None,
-                stale_if_error_secs: None,
-                use_stale: None,
-                background_update: None,
-                lock_timeout_secs: None,
-                lock_age_secs: None,
-                min_uses: None,
-                ignore_headers: None,
-                range_requests: None,
-                slice_size_bytes: None,
-                convert_head: Some(false),
-            }),
-            matcher: MatcherConfig::Prefix("/assets".to_string()),
-            handler: HandlerConfig::Proxy {
-                upstream: "backend".to_string(),
-                preserve_host: None,
-                strip_prefix: None,
-                proxy_set_headers: std::collections::HashMap::new(),
-            },
-            grpc_service: None,
-            grpc_method: None,
-            allow_cidrs: Vec::new(),
-            deny_cidrs: Vec::new(),
-            requests_per_sec: None,
-            burst: None,
-            allow_early_data: None,
-            request_buffering: None,
-            response_buffering: None,
-            compression: None,
-            compression_min_bytes: None,
-            compression_content_types: None,
-            streaming_response_idle_timeout_secs: None,
-        }],
-        servers: Vec::new(),
-    };
-
-    let snapshot =
-        compile_with_base(config, base_dir.path()).expect("p2 cache policy should compile");
-    assert!(!snapshot.cache_zones["default"].shared_index);
-    assert!(
-        !snapshot.default_vhost.routes[0]
-            .cache
-            .as_ref()
-            .expect("route cache policy should compile")
-            .convert_head
-    );
-}
-
-#[test]
-fn compile_cache_policy_supports_p3_slice_controls() {
-    let base_dir = temp_base_dir("rginx-cache-compile-p3");
-    let config = Config {
-        cache_zones: vec![CacheZoneConfig {
-            name: "default".to_string(),
-            path: "cache/default".to_string(),
-            max_size_bytes: Some(1024 * 1024),
-            inactive_secs: Some(120),
-            default_ttl_secs: Some(30),
-            max_entry_bytes: Some(1024),
-            path_levels: None,
-            loader_batch_entries: None,
-            loader_sleep_millis: None,
-            manager_batch_entries: None,
-            manager_sleep_millis: None,
-            inactive_cleanup_interval_secs: None,
-            shared_index: None,
-        }],
-        runtime: RuntimeConfig {
-            shutdown_timeout_secs: 10,
-            worker_threads: None,
-            accept_workers: None,
-        },
-        listeners: Vec::new(),
-        server: ServerConfig {
-            listen: Some("127.0.0.1:8080".to_string()),
-            server_header: None,
-            proxy_protocol: None,
-            default_certificate: None,
-            server_names: Vec::new(),
-            trusted_proxies: Vec::new(),
-            client_ip_header: None,
-            keep_alive: None,
-            max_headers: None,
-            max_request_body_bytes: None,
-            max_connections: None,
-            header_read_timeout_secs: None,
-            request_body_read_timeout_secs: None,
-            response_write_timeout_secs: None,
-            access_log_format: None,
-            tls: None,
-            http3: None,
-        },
-        upstreams: vec![UpstreamConfig {
-            name: "backend".to_string(),
-            peers: vec![UpstreamPeerConfig {
-                url: "http://127.0.0.1:9000".to_string(),
-                weight: 1,
-                backup: false,
-            }],
-            tls: None,
-            dns: None,
-            protocol: UpstreamProtocolConfig::Auto,
-            load_balance: UpstreamLoadBalanceConfig::RoundRobin,
-            server_name: None,
-            server_name_override: None,
-            request_timeout_secs: None,
-            connect_timeout_secs: None,
-            read_timeout_secs: None,
-            write_timeout_secs: None,
-            idle_timeout_secs: None,
-            pool_idle_timeout_secs: None,
-            pool_max_idle_per_host: None,
-            tcp_keepalive_secs: None,
-            tcp_nodelay: None,
-            http2_keep_alive_interval_secs: None,
-            http2_keep_alive_timeout_secs: None,
-            http2_keep_alive_while_idle: None,
-            max_replayable_request_body_bytes: None,
-            unhealthy_after_failures: None,
-            unhealthy_cooldown_secs: None,
-            health_check_path: None,
-            health_check_grpc_service: None,
-            health_check_interval_secs: None,
-            health_check_timeout_secs: None,
-            healthy_successes_required: None,
-        }],
-        locations: vec![LocationConfig {
-            cache: Some(CacheRouteConfig {
-                zone: "default".to_string(),
-                methods: Some(vec!["GET".to_string()]),
-                statuses: Some(vec![200]),
-                ttl_secs_by_status: None,
-                key: Some("{scheme}:{host}:{uri}".to_string()),
-                cache_bypass: None,
-                no_cache: None,
-                stale_if_error_secs: None,
-                use_stale: None,
-                background_update: None,
-                lock_timeout_secs: None,
-                lock_age_secs: None,
-                min_uses: None,
-                ignore_headers: None,
-                range_requests: Some(crate::model::CacheRangeRequestPolicyConfig::Cache),
-                slice_size_bytes: Some(128),
-                convert_head: None,
-            }),
-            matcher: MatcherConfig::Prefix("/assets".to_string()),
-            handler: HandlerConfig::Proxy {
-                upstream: "backend".to_string(),
-                preserve_host: None,
-                strip_prefix: None,
-                proxy_set_headers: std::collections::HashMap::new(),
-            },
-            grpc_service: None,
-            grpc_method: None,
-            allow_cidrs: Vec::new(),
-            deny_cidrs: Vec::new(),
-            requests_per_sec: None,
-            burst: None,
-            allow_early_data: None,
-            request_buffering: None,
-            response_buffering: None,
-            compression: None,
-            compression_min_bytes: None,
-            compression_content_types: None,
-            streaming_response_idle_timeout_secs: None,
-        }],
-        servers: Vec::new(),
-    };
-
-    let snapshot =
-        compile_with_base(config, base_dir.path()).expect("p3 cache policy should compile");
-    assert_eq!(
-        snapshot.default_vhost.routes[0]
-            .cache
-            .as_ref()
-            .expect("route cache policy should compile")
-            .slice_size_bytes,
-        Some(128)
-    );
-}
