+++ crates/rginx-http/src/cache/store.rs
@@ -1,11 +1,13 @@
 use std::sync::{Arc, Mutex};
 use std::time::{Duration, SystemTime};
 
+use bytes::Bytes;
 use http::StatusCode;
+use http::header::HeaderMap;
 use http_body_util::BodyExt;
 use tokio::fs;
 
-use crate::handler::HttpResponse;
+use crate::handler::{HttpResponse, full_body};
 
 use super::entry::{
     CacheMetadataInput, build_cached_response_for_request, cache_key_hash, cache_metadata,
@@ -23,6 +25,7 @@ use super::{
 
 mod helpers;
 mod maintenance;
+mod revalidate;
 
 use helpers::{
     cache_metadata_input, cache_vary_values, freshness_is_cacheable, merge_not_modified_headers,
@@ -32,6 +35,7 @@ pub(super) use maintenance::{
     cleanup_inactive_entries_in_zone, eviction_candidates, lock_index, purge_zone_entries,
     record_cache_admission_attempt, remove_index_entry,
 };
+pub(in crate::cache) use revalidate::refresh_not_modified_response;
 
 pub(crate) struct CacheStoreError {
     source: Box<dyn std::error::Error + Send + Sync>,
@@ -79,16 +83,8 @@ pub(super) async fn store_response(
             return Err(CacheStoreError { source: error });
         }
     };
-    let downstream_response = || {
-        finalize_response_for_request(
-            parts.status,
-            &parts.headers,
-            collected.as_ref(),
-            &context.request,
-            &context.policy,
-        )
-        .map_err(|error| CacheStoreError { source: Box::new(error) })
-    };
+    let downstream_response =
+        || finalize_downstream_response(&context, parts.status, &parts.headers, collected.as_ref());
 
     if !storable || no_cache || collected.len() > context.zone.config.max_entry_bytes {
         return downstream_response();
@@ -162,149 +158,6 @@ pub(super) async fn store_response(
     downstream_response()
 }
 
-pub(super) async fn refresh_not_modified_response(
-    context: CacheStoreContext,
-    response: HttpResponse,
-) -> std::result::Result<HttpResponse, CacheStoreError> {
-    let Some(cached_entry) = context.cached_entry.clone() else {
-        context.zone.record_write_error();
-        return Err(CacheStoreError {
-            source: Box::new(std::io::Error::other("missing cached entry for 304 revalidation")),
-        });
-    };
-    let Some(cached_metadata) = context.cached_metadata.clone() else {
-        context.zone.record_write_error();
-        return Err(CacheStoreError {
-            source: Box::new(std::io::Error::other("missing cached metadata for 304 revalidation")),
-        });
-    };
-
-    let cached_headers = cached_metadata
-        .headers_map()
-        .map_err(|error| CacheStoreError { source: Box::new(error) })?;
-    let merged_headers = merge_not_modified_headers(&cached_headers, response.headers());
-    let now = unix_time_ms(SystemTime::now());
-    let cached_status = StatusCode::from_u16(cached_metadata.status).unwrap_or(StatusCode::OK);
-    let paths = cache_paths_for_zone(context.zone.config.as_ref(), &cached_entry.hash);
-    if response_no_cache(&context, cached_status) {
-        let response_metadata = cache_metadata(
-            cached_metadata.key.clone(),
-            cached_status,
-            &merged_headers,
-            CacheMetadataInput {
-                base_key: cached_metadata.base_key.clone(),
-                vary: cached_entry.vary.clone(),
-                stored_at_unix_ms: cached_metadata.stored_at_unix_ms,
-                expires_at_unix_ms: cached_metadata.expires_at_unix_ms,
-                stale_if_error_until_unix_ms: cached_metadata.stale_if_error_until_unix_ms,
-                stale_while_revalidate_until_unix_ms: cached_metadata
-                    .stale_while_revalidate_until_unix_ms,
-                must_revalidate: cached_metadata.must_revalidate,
-                body_size_bytes: cached_metadata.body_size_bytes,
-            },
-        );
-        let refreshed = {
-            let _io_guard = context.zone.io_lock.lock().await;
-            build_cached_response_for_request(
-                &paths.body,
-                &response_metadata,
-                &context.request,
-                &context.policy,
-                context.read_cached_body,
-            )
-            .await
-        };
-        return refreshed
-            .map(|response| with_cache_status(response, CacheStatus::Revalidated))
-            .map_err(|error| CacheStoreError { source: Box::new(error) });
-    }
-
-    let freshness = response_freshness(&context, cached_status, &merged_headers);
-    let vary = cache_vary_values(&context, &context.request, &merged_headers);
-    let final_key = cache_variant_key(&context.base_key, &vary);
-    let metadata = cache_metadata(
-        final_key.clone(),
-        cached_status,
-        &merged_headers,
-        cache_metadata_input(
-            &context.base_key,
-            vary.clone(),
-            now,
-            &freshness,
-            cached_metadata.body_size_bytes,
-        ),
-    );
-    if !response_is_storable_with_size(
-        &context,
-        cached_status,
-        &merged_headers,
-        ResponseBodySize::exact(cached_metadata.body_size_bytes),
-    ) || !freshness_is_cacheable(&freshness)
-        || final_key != context.key
-    {
-        let refreshed = {
-            let _io_guard = context.zone.io_lock.lock().await;
-            build_cached_response_for_request(
-                &paths.body,
-                &metadata,
-                &context.request,
-                &context.policy,
-                context.read_cached_body,
-            )
-            .await
-            .map_err(|error| CacheStoreError { source: Box::new(error) })
-        };
-        remove_index_entry(&context.zone, &context.key);
-        {
-            let _io_guard = context.zone.io_lock.lock().await;
-            let _ = fs::remove_file(&paths.metadata).await;
-            let _ = fs::remove_file(&paths.body).await;
-        }
-        persist_zone_shared_index(&context.zone).await;
-        context.zone.record_revalidated();
-        return refreshed.map(|response| with_cache_status(response, CacheStatus::Revalidated));
-    }
-    {
-        let _io_guard = context.zone.io_lock.lock().await;
-        write_cache_metadata(&paths, &metadata)
-            .await
-            .map_err(|error| CacheStoreError { source: Box::new(error) })?;
-    }
-    context.zone.record_write_success();
-    context.zone.record_revalidated();
-    update_index_after_store(
-        &context.zone,
-        context.key,
-        CacheIndexEntry {
-            hash: cached_entry.hash.clone(),
-            base_key: context.base_key.clone(),
-            vary,
-            body_size_bytes: metadata.body_size_bytes,
-            expires_at_unix_ms: metadata.expires_at_unix_ms,
-            stale_if_error_until_unix_ms: metadata.stale_if_error_until_unix_ms,
-            stale_while_revalidate_until_unix_ms: metadata.stale_while_revalidate_until_unix_ms,
-            must_revalidate: metadata.must_revalidate,
-            last_access_unix_ms: now,
-        },
-        None,
-    )
-    .await;
-
-    let refreshed = {
-        let _io_guard = context.zone.io_lock.lock().await;
-        build_cached_response_for_request(
-            &paths.body,
-            &metadata,
-            &context.request,
-            &context.policy,
-            context.read_cached_body,
-        )
-        .await
-        .map_err(|error| CacheStoreError { source: Box::new(error) })?
-    };
-    Ok(with_cache_status(refreshed, CacheStatus::Revalidated))
-}
-
 async fn update_index_after_store(
     zone: &Arc<CacheZoneRuntime>,
     key: String,
@@ -316,4 +169,46 @@ async fn update_index_after_store(
 
 pub(super) fn duration_to_ms(duration: Duration) -> u64 {
     duration.as_millis().min(u128::from(u64::MAX)) as u64
+}
+
+fn finalize_downstream_response(
+    context: &CacheStoreContext,
+    status: StatusCode,
+    headers: &HeaderMap,
+    body: &[u8],
+) -> std::result::Result<HttpResponse, CacheStoreError> {
+    if super::request::cacheable_range_request(&context.request, &context.policy)
+        .is_some_and(|range| range.needs_downstream_trimming())
+        && !downstream_range_trim_compatible(context, status, headers)
+    {
+        return build_response(status, headers, body);
+    }
+
+    finalize_response_for_request(status, headers, body, &context.request, &context.policy)
+        .map_err(|error| CacheStoreError { source: Box::new(error) })
+}
+
+fn downstream_range_trim_compatible(
+    context: &CacheStoreContext,
+    status: StatusCode,
+    headers: &HeaderMap,
+) -> bool {
+    status == StatusCode::PARTIAL_CONTENT
+        && super::request::response_content_range_matches_request(
+            &context.request,
+            &context.policy,
+            headers,
+        )
+}
+
+fn build_response(
+    status: StatusCode,
+    headers: &HeaderMap,
+    body: &[u8],
+) -> std::result::Result<HttpResponse, CacheStoreError> {
+    let mut response = http::Response::builder().status(status);
+    *response.headers_mut().expect("response builder should expose headers") = headers.clone();
+    response.body(full_body(Bytes::copy_from_slice(body))).map_err(|error| CacheStoreError {
+        source: Box::new(std::io::Error::other(error.to_string())),
+    })
 }
