+++ crates/rginx-http/src/cache/tests/lookup/keys.rs
@@ -277,145 +277,3 @@ fn multiple_range_headers_bypass_cache_when_range_caching_is_enabled() {
 
     assert!(cache_request_bypass(&request, &policy));
 }
-
-#[test]
-fn head_cache_key_is_separated_when_convert_head_is_disabled() {
-    let mut policy = RouteCachePolicy {
-        zone: "default".to_string(),
-        methods: vec![Method::GET, Method::HEAD],
-        statuses: vec![StatusCode::OK],
-        ttl_by_status: Vec::new(),
-        key: rginx_core::CacheKeyTemplate::parse("{scheme}:{host}:{uri}")
-            .expect("key should parse"),
-        cache_bypass: None,
-        no_cache: None,
-        stale_if_error: None,
-        use_stale: Vec::new(),
-        background_update: false,
-        lock_timeout: Duration::from_secs(5),
-        lock_age: Duration::from_secs(5),
-        min_uses: 1,
-        ignore_headers: Vec::new(),
-        range_requests: rginx_core::CacheRangeRequestPolicy::Bypass,
-        slice_size_bytes: None,
-        convert_head: false,
-    };
-    let get_request = Request::builder()
-        .method(Method::GET)
-        .uri("/head-key")
-        .header("host", "example.com")
-        .body(full_body(Bytes::new()))
-        .expect("request should build");
-    let head_request = Request::builder()
-        .method(Method::HEAD)
-        .uri("/head-key")
-        .header("host", "example.com")
-        .body(full_body(Bytes::new()))
-        .expect("request should build");
-
-    let get_key = render_cache_key(
-        get_request.method(),
-        get_request.uri(),
-        get_request.headers(),
-        "https",
-        &policy,
-    );
-    let head_key = render_cache_key(
-        head_request.method(),
-        head_request.uri(),
-        head_request.headers(),
-        "https",
-        &policy,
-    );
-
-    assert_eq!(get_key, "https:example.com:/head-key|cache-method:GET");
-    assert_eq!(head_key, "https:example.com:/head-key|cache-method:HEAD");
-
-    policy.convert_head = true;
-    let shared_head_key = render_cache_key(
-        head_request.method(),
-        head_request.uri(),
-        head_request.headers(),
-        "https",
-        &policy,
-    );
-    assert_eq!(shared_head_key, "https:example.com:/head-key");
-}
-
-#[test]
-fn cache_key_reuses_slice_for_subranges_within_same_slice() {
-    let policy = RouteCachePolicy {
-        zone: "default".to_string(),
-        methods: vec![Method::GET],
-        statuses: vec![StatusCode::OK],
-        ttl_by_status: Vec::new(),
-        key: rginx_core::CacheKeyTemplate::parse("{scheme}:{host}:{uri}")
-            .expect("key should parse"),
-        cache_bypass: None,
-        no_cache: None,
-        stale_if_error: None,
-        use_stale: Vec::new(),
-        background_update: false,
-        lock_timeout: Duration::from_secs(5),
-        lock_age: Duration::from_secs(5),
-        min_uses: 1,
-        ignore_headers: Vec::new(),
-        range_requests: rginx_core::CacheRangeRequestPolicy::Cache,
-        slice_size_bytes: Some(128),
-        convert_head: true,
-    };
-    let first = Request::builder()
-        .method(Method::GET)
-        .uri("/video.mp4")
-        .header("host", "example.com")
-        .header(http::header::RANGE, "bytes=2-4")
-        .body(full_body(Bytes::new()))
-        .expect("request should build");
-    let second = Request::builder()
-        .method(Method::GET)
-        .uri("/video.mp4")
-        .header("host", "example.com")
-        .header(http::header::RANGE, "bytes=5-6")
-        .body(full_body(Bytes::new()))
-        .expect("request should build");
-
-    assert_eq!(
-        render_cache_key(first.method(), first.uri(), first.headers(), "https", &policy),
-        "https:example.com:/video.mp4|range:0-127"
-    );
-    assert_eq!(
-        render_cache_key(second.method(), second.uri(), second.headers(), "https", &policy),
-        "https:example.com:/video.mp4|range:0-127"
-    );
-}
-
-#[test]
-fn cross_slice_range_bypasses_when_slice_size_is_configured() {
-    let policy = RouteCachePolicy {
-        zone: "default".to_string(),
-        methods: vec![Method::GET],
-        statuses: vec![StatusCode::OK],
-        ttl_by_status: Vec::new(),
-        key: rginx_core::CacheKeyTemplate::parse("{uri}").expect("key should parse"),
-        cache_bypass: None,
-        no_cache: None,
-        stale_if_error: None,
-        use_stale: Vec::new(),
-        background_update: false,
-        lock_timeout: Duration::from_secs(5),
-        lock_age: Duration::from_secs(5),
-        min_uses: 1,
-        ignore_headers: Vec::new(),
-        range_requests: rginx_core::CacheRangeRequestPolicy::Cache,
-        slice_size_bytes: Some(8),
-        convert_head: true,
-    };
-    let request = Request::builder()
-        .method(Method::GET)
-        .uri("/video.mp4")
-        .header(http::header::RANGE, "bytes=6-9")
-        .body(full_body(Bytes::new()))
-        .expect("request should build");
-
-    assert!(cache_request_bypass(&CacheRequest::from_request(&request), &policy));
-}
