+++ crates/rginx-http/src/cache/tests/storage_p3.rs
@@ -63,3 +63,42 @@ async fn sliced_range_requests_rewrite_upstream_range_and_trim_hits() {
         _ => panic!("subrange within the same slice should hit the cached slice"),
     }
 }
+
+#[tokio::test]
+async fn sliced_range_requests_fallback_to_passthrough_when_origin_ignores_range() {
+    let temp = tempfile::tempdir().expect("cache temp dir should exist");
+    let manager = test_manager(temp.path().to_path_buf(), 1024);
+    let mut policy = test_policy();
+    policy.range_requests = rginx_core::CacheRangeRequestPolicy::Cache;
+    policy.slice_size_bytes = Some(8);
+
+    let request = Request::builder()
+        .method(Method::GET)
+        .uri("/slice")
+        .header("host", "example.com")
+        .header(RANGE, "bytes=2-4")
+        .body(full_body(Bytes::new()))
+        .expect("request should build");
+    let context = match manager.lookup(CacheRequest::from_request(&request), "https", &policy).await
+    {
+        CacheLookup::Miss(context) => *context,
+        _ => panic!("slice request should miss before storing"),
+    };
+
+    let upstream_response = Response::builder()
+        .status(StatusCode::OK)
+        .header(CONTENT_LENGTH, "8")
+        .body(full_body("abcdefgh"))
+        .expect("response should build");
+    let downstream = manager.store_response(context, upstream_response).await;
+    assert_eq!(downstream.status(), StatusCode::OK);
+    assert_eq!(downstream.headers().get(CACHE_STATUS_HEADER).unwrap(), "MISS");
+    assert!(downstream.headers().get(CONTENT_RANGE).is_none());
+    let body = downstream.into_body().collect().await.unwrap().to_bytes();
+    assert_eq!(body.as_ref(), b"abcdefgh");
+
+    assert!(matches!(
+        manager.lookup(CacheRequest::from_request(&request), "https", &policy).await,
+        CacheLookup::Miss(_)
+    ));
+}
